+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## (UNRELEASED)
+- Add name and size tag to ActiveCampaign subscribers
+- Add subscribers to Customer list in ActiveCampaign
 - Background task install jobs (ActiveCampaign Subscribe + Slack Notify)
 - Bump default send_delay to 1 week